refactor for wallets
