+++ app/build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'com.android.application'
     id 'kotlin-android'
     id 'kotlin-kapt'
+    id 'com.google.gms.google-services'
 }
 
 android {
@@ -77,5 +78,16 @@ dependencies {
     // room
     implementation "androidx.room:room-runtime:2.3.0"
     kapt "androidx.room:room-compiler:2.3.0"
+    implementation "androidx.room:room-rxjava2:2.3.0"
+
+    //rxjava
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.21'
+    runtimeOnly 'io.reactivex.rxjava2:rxkotlin:2.4.0'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+
+    //firebase
+    implementation 'com.google.firebase:firebase-firestore:23.0.3'
+
+
 
 }
