+++ app/src/main/java/com/leushinilya/loftcoin/AppComponent.kt
@@ -2,10 +2,11 @@ package com.leushinilya.loftcoin
 
 import android.app.Application
 import android.content.Context
-import com.leushinilya.loftcoin.data.db.CoinsRepoDB
-import com.leushinilya.loftcoin.data.remote.CmcAPI
+import com.leushinilya.loftcoin.data.rates.db.CoinsRepoDB
+import com.leushinilya.loftcoin.data.rates.remote.CmcAPI
 import com.leushinilya.loftcoin.data.DataModule
-import com.leushinilya.loftcoin.data.db.LoftDatabase
+import com.leushinilya.loftcoin.data.rates.db.LoftDatabase
+import com.leushinilya.loftcoin.data.wallets.WalletsRepo
 import com.leushinilya.loftcoin.ui.main.rates.RatesViewModel
 import com.leushinilya.loftcoin.utils.RatesViewModelFactory
 import com.leushinilya.loftcoin.utils.ViewModelModule
@@ -26,10 +27,14 @@ abstract class AppComponent {
         abstract fun build(): AppComponent
     }
 
+//    provide
     abstract fun context(): Context
     abstract fun cmcAPI(): CmcAPI
     abstract fun database(): LoftDatabase
-    abstract fun coinsRepo(): CoinsRepoDB
+    abstract fun coinsRepoDB(): CoinsRepoDB
+    abstract fun walletsRepo(): WalletsRepo
+
+//    inject
     abstract fun inject(ratesViewModel: RatesViewModel)
