+++ app/src/main/java/com/leushinilya/loftcoin/data/rates/db/CoinsDAO.java
@@ -1,13 +1,11 @@
-package com.leushinilya.loftcoin.data.db;
+package com.leushinilya.loftcoin.data.rates.db;
 
 import static androidx.room.OnConflictStrategy.REPLACE;
 
 import androidx.annotation.WorkerThread;
-import androidx.lifecycle.LiveData;
 import androidx.room.Dao;
 import androidx.room.Insert;
 import androidx.room.Query;
-import androidx.room.Room;
 
 import java.util.List;
