+++ app/src/main/java/com/leushinilya/loftcoin/data/rates/remote/Listings.java
@@ -1,7 +1,6 @@
-package com.leushinilya.loftcoin.data.remote;
+package com.leushinilya.loftcoin.data.rates.remote;
 
 import com.google.gson.annotations.SerializedName;
-import com.leushinilya.loftcoin.data.remote.CmcCoin;
 
 import java.util.List;
