+++ app/src/main/java/com/leushinilya/loftcoin/data/wallets/Transaction.kt
@@ -1,0 +1,6 @@
+package com.leushinilya.loftcoin.data.wallets
+
+import com.leushinilya.loftcoin.data.rates.Coin
+import java.util.*
+
+data class Transaction(val uid:String, val coin: Coin, val amount: Double, val createdAt: Date)
