+++ app/src/main/java/com/leushinilya/loftcoin/data/wallets/Wallet.kt
@@ -1,0 +1,6 @@
+package com.leushinilya.loftcoin.data.wallets
+
+import com.leushinilya.loftcoin.data.rates.Coin
+
+data class Wallet (val uid: String, val coin: Coin, val balance: Double){
+}
