+++ app/src/main/java/com/leushinilya/loftcoin/utils/RatesViewModelFactory.kt
@@ -2,7 +2,7 @@ package com.leushinilya.loftcoin.utils
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.leushinilya.loftcoin.data.remote.CmcAPI
+import com.leushinilya.loftcoin.data.rates.remote.CmcAPI
 import com.leushinilya.loftcoin.ui.main.rates.RatesViewModel
 import javax.inject.Inject
